+++ src/pyramid_bimt/views/misc.py
@@ -4,7 +4,6 @@
 from pyramid.httpexceptions import exception_response
 from pyramid.view import view_config
 from datetime import datetime
-from pyramid.response import Response
 
 
 @view_config(
@@ -18,14 +17,15 @@ def raise_http_error(request):
 @view_config(
     route_name='raise_js_error',
     permission='admin',
+    layout='default',
+    renderer='pyramid_bimt:templates/form.pt',
 )
 def raise_js_error(request):
-    body = """<script type="text/javascript">
+    return {
+        'title': 'JS error',
+        'form': """<script type="text/javascript">
       throw new Error('[{now}] Error test.');
     </script>
 
     <p>[{now}] Error test.</p>""".format(now=datetime.utcnow())
-    return Response(
-        content_type='text/html',
-        body=body
-    )
+    }